Improve error message for I18n.tp (#30)
* Improve error message for I18n.tp

We've been getting errors like this one:

```
You must have a `count` property
```

But the code is this:

```js
let collLength = collection?.length
i18n.tp('foo.bar', {
    count: collLength,
  })
```

The error is a bit misleading, because the property `count` **is defined**, but the problem is that it's null. 

This commit fixes the error message so it's a bit more clear about what might be happening.